refactor(frontend): share thinking level control
