+++ apps/frontend/src/modules/chat-session/components/ThinkingLevelSelect/index.ts
@@ -1,1 +1,2 @@
+export {useThinkingLevel} from './hooks/useThinkingLevel.js';
 export {ThinkingLevelSelect} from './ThinkingLevelSelect.js';
