+++ apps/frontend/src/modules/chat-session/index.ts
@@ -5,6 +5,10 @@ export {ChatInput} from './components/ChatInput/index.js';
 export {InfoBar} from './components/InfoBar/index.js';
 export {SessionSidebar} from './components/SessionSidebar/index.js';
 export {StreamingMessageDisplay} from './components/StreamingMessageDisplay/index.js';
+export {
+  ThinkingLevelSelect,
+  useThinkingLevel,
+} from './components/ThinkingLevelSelect/index.js';
 export {TitleBarView} from './components/TitleBar/index.js';
 
 // Contexts (providers + values)
